input_as_shape option for Filler ops
Summary: I couldn't find a way to fill a tensor with a shape provided at runtime, so I added an input_as_shape option to the filler ops. When input_as_shape is true, the input can be used to directly provide the shape of the output (this is different from the default behavior, where the output is reshaped like the input). For example if the input contains [2, 3], the output will have shape [2, 3]. Let me know if you see a simpler way :)

Reviewed By: Yangqing

Differential Revision: D4276872

fbshipit-source-id: 095e995d8bf302152765bd51c405185ef9952212