add HTraceAsyncDAGNet
Summary:
This diff adds HTraceAsyncDAGNet, which is basically the async_dag version of HTraceDAGNet. Similar to HTraceDAGNet, we can use HTraceAsyncDAGNet by setting the net type to `htrace_async_dag`.

For now, we only track iteration spans and do not go deeper (operators, gpu kernels, etc.) because due to the implementation of AsyncDAGNet, applying HTrace is much more intrusive compared to HTraceDAGNet. Creating spans for operators for HTraceAsyncDAGNet is a future task.

This diff also adds a minor change in the TARGETS file so that `htrace_dag`, `htrace_async_dag`, and `prof_dag` are all accessible via one rule.

Differential Revision: D4351587

fbshipit-source-id: 1a4075a9a5efdfafb828a81b663cc731858f7307