Don't continue running operators after failure
Summary:
This brings back DAGNet up to parity with SimpleNet, where
execution stops as expected after an operator fails. For the DAGNet
it's more involved, since we have to deal with all worker threads
stopping execution. Because the job queue may still hold an arbitrary
number of chains to execute, this diff explicitly closes it down,
waits for all workers to terminate, and resets the job queue, upon
seeing a failure.

Reviewed By: akyrola

Differential Revision: D5232955

fbshipit-source-id: 4dac3c3ed6e5c2ebd07473b0f8be2b02c28978e9