Avoid grabbing DeviceGuard in at::empty when possible (#13785)
Summary:
Changed at::empty to allocate the correct amount of memory instead of
"allocate 0 memory and then resize it to the necessary size".

This leads to a 300 ns speedup for at::empty for a cuda tensor of size (64, 2048).
(1790ns -> 1460ns for at::empty).

Also does the following:
Removes DeviceGuards for:
- empty_* functions that end up calling functions that already have a
  DeviceGuard
- t(), which gets called a lot in LSTMs,
- Remove one of the two DeviceGuard that at::empty(...) uses. It only
  needs one for correctness, the other comes from the resize_
  implementation.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/13785

Reviewed By: ezyang

Differential Revision: D13004938

Pulled By: zou3519

fbshipit-source-id: f45b7e6abe06c05d1f81cc53e190c7bab6d1c116