Deprecate unordered_map and vector in IValues (#21712)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/21712

Warn when people use unordered_map or vector with IValues. These APIs are deprecated.
The unordered_map API is slow because it requires copying the whole map.
The vector API is slow for some types (e.g. std::string) because for them it also requires copying the whole map.
Also, the vector API would get slow for all types if we decide to switch to SmallVector.

Differential Revision: D15792428

fbshipit-source-id: 1b72406b3a8d56521c862858c9f0ed01e56f2757