Make scatter/gather arguments optional (#25575)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25575

For both scatter and gather, only the source and destination rank,
respectively, need to supply a list of tensors. The `scatter_list` and
`gather_list` arguments were mandatory, however, and this has resulted
in some confusion. This commit makes both the `scatter_list` and
`gather_list`, and the `src` and `dst` arguments optional.

Closes #25463.

Test Plan: Imported from OSS

Differential Revision: D17164253

fbshipit-source-id: a16bc208c87a1c96163c1a86d4a7ca8634a26f95