Delete TEST_NAMEDTENSOR; run named tensor tests on all CIs (#27760)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27760

There's nothing special about the named tensor tests that requires that
they be run in their own CI job. In this PR we delete the
TEST_NAMEDTENSOR flag that hides named tensor tests from regular jobs.
In the future, we'll delete the named tensor CI job so that we do not
duplicate signals.

Test Plan: - wait for CI

Differential Revision: D17882262

Pulled By: zou3519

fbshipit-source-id: f90c71cb939e53b8ea23f7e2ab95a5c41b8be0e3