Add note about when to install typing package (#29103)
Summary:
Was just trying to build pytorch from source and had a small hiccup because the instructions say to `conda install typing`. Because `typing` is a built-in module in recent Python 3 versions, conda interpreted that to mean that I want Python 2. So I added a note to the docs.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29103

Differential Revision: D18294139

Pulled By: soumith

fbshipit-source-id: 621a2f62ebe870520197baec8f8bcdc1a0c57de9