Support process_group_agent "sending to itself" (#29253)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29253

Some operations can be simpler if a worker can send an rpc to itself.
The main reason for not doing previous was that Gloo doesn't support
self-sending.

That said, this changes the process_group_agent to skip the assert
check, and simply enqueue the rpc message in its receiving queue.
ghstack-source-id: 93518076

Test Plan: buck test mode/dev-nosan caffe2/test/...

Differential Revision: D18339715

fbshipit-source-id: 08ade40e81da378b003a550c898a726e99d50e34