omit constexpr with nvcc on clang (#32149)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32149

This is an attempt at clarifying some of the preprocessor boolean logic that was getting more and more complicated. The previous logic used constexpr with nvcc on clang; which we were getting compiler failures on in ovrsource with mode/linux/* (based on platform007).

Test Plan:
ovrsource xplat/caffe2 compiles
fbsource sandcastle green

Differential Revision: D19385409

fbshipit-source-id: 60a02bae9854388b87510afdd927709673a6c313