[pytorch] blas gemm fix for k=0 (#33419)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33419

These conditions are for the specific implementation, the fallback implementation works without these checks. So use that if any of these checks isn't true.
ghstack-source-id: 98836075

Test Plan: Previously got error for special case where k=0 which has gone. The error was in some complicated autograd, and I'm not sure how and where an simple regression test should be added.

Differential Revision: D19941103

fbshipit-source-id: e1c85d1e75744b1c51ad9b71c7b3211af3c5bcc6