[RpcAgent] Metrics for current num active/async rpc calls. (#34398)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34398

As part of PR 34109, it was suggested that we track the number of outstanding
async calls for RPC DebugInfo, particularly if we move towards using
at::launch() threads on occasion for continuations.

This particular aspect of the change was distinct from the main purpose of the
diff, and started getting bigger, so split this functionality out as a separate diff.
For completeness, we track client_active_calls, server_active_calls,
server_active_async_calls, and write some very basic unittest coverage.
ghstack-source-id: 99708836

Test Plan: buck test mode/dev-nosan caffe2/torch/fb/distributed/thriftRpcBackend/...

Differential Revision: D20314994

fbshipit-source-id: 2f7c75d5c511b27ed0c09c7b8a67b6fb49df31a5