Remove AllGatherTestWithTimeout (#41945)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41945

This test previously did a thread sleep before launching the allgather operation, and then waited on the work object. Since the sleep was done before the work object was created, it did not affect the allgather call, and thus, did not test work-level timeouts as intended.

I am removing this test for now. In the future we can add this test back, but would need to somehow inject a `cudaSleep` call before the  allgather (so the collective operation itself is delayed). This may require overriding the `ProcessGroupNCCL::collective`, so it's a bit more heavy-weight.

In the meantime, we can remove this test - work-level timeouts are still thoroughly tested with Gloo.
ghstack-source-id: 108370178

Test Plan: Ran ProcessGroupNCCL tests on devGPU

Reviewed By: jiayisuse

Differential Revision: D22702291

fbshipit-source-id: a36ac3d83abfab6351c0476046a2f3b04a80c44d