inline `has` function for DispatchKeySet (#49191)
Summary:
inlines `has` function for DispatchKeySet, that is frequently used in TensorImpl in calls such as `is_sparse`, `is_cuda` etc.
This increases `empty` instruction count (1853228 -> 1937428) without appreciable effect on runtime, and noticeably reduces instruction counts for `copy_` and friends that have to rely on `is_sparse`, `is_cuda` and the like a lot to decide which path to take (3269114 -> 2634114).

Pull Request resolved: https://github.com/pytorch/pytorch/pull/49191

Reviewed By: H-Huang

Differential Revision: D25483011

Pulled By: ngimel

fbshipit-source-id: 2f3ab83e2c836a726b9284ffc50d6ecf3701aada