[quant] Backend string for the quantized types (#49965)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49965

Without this checking the type of the quantized tensor using `type` would throw an error.

After this PR running the `type(qx)`, where `qx` is a quantized tensor would show something like `torch.quantized.QUInt8`.

Test Plan: Not needed -- this is just a string description for the quantized tensors

Differential Revision: D25731594

Reviewed By: ezyang

Pulled By: z-a-f

fbshipit-source-id: 942fdf89a1c50895249989c7203f2e7cc00df4c6