ns for fx: remove subgraphs from user facing API (#52928)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52928

Changes the user facing API of `prepare_single_model_output` to
require a list of nodes instead of a list of subgraphs. This ensures
that how we define a subgraph is an implementation detail and is
not exposed to the user, keeping the eng cost of updating this
implementation later low.

Test Plan:
```
python test/test_quantization.py TestFXNumericSuiteCoreAPIs
```

Imported from OSS

Reviewed By: raghuramank100

Differential Revision: D26693471

fbshipit-source-id: 67c2feb844556225e36f8d6d4023246939bcb445