Remove ci-all and release branches running scheduled tests (#53069)
Summary:
The previous code allowed these tests to run every four hours on certain ci-all branches...which is really bad and resource intensive. This code removes that, but then disallows the 11.2 and 9.2 tests to be run on ci-all branches.

To debug CUDA 11.2 or 9.2 tests, one must now manually change the config to allow for them. (Look at https://github.com/pytorch/pytorch/issues/51888 and https://github.com/pytorch/pytorch/issues/51598 for examples of how to do that.)

Pull Request resolved: https://github.com/pytorch/pytorch/pull/53069

Reviewed By: H-Huang

Differential Revision: D26739738

Pulled By: janeyx99

fbshipit-source-id: 7577b9b2e876bac0e4e868ce2a1f3ffdb6aca597