[libkineto] Log CUPTI errors on libkineto initialization
Summary: When libkineto is initialized from the PyTorch Profiler, if it fails we will not know why because errors are not reported. Reporting errors is not always safe, e.g. if init happens from static initialization or a dlopen library constructor function, so add a flag to specify whether to log.

Test Plan: Testing in PyTorch OSS build.

Reviewed By: chaekit

Differential Revision: D26927500

fbshipit-source-id: 2a78005239a5fcbe7e1de82e5405f04e07000fa8