[PyPer] Port c2 add to pt (#54229)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54229

Because caffe2 add uses Eigen for add with broadcasting which is not well supported by OSS PyTorch, it's easier to just keep the `c2_add_out` internal for now. Caffe2 does use mkl add when the input dims of A and B are the same and there is no broadcasting needed.

Reviewed By: bertmaher

Differential Revision: D27036279

fbshipit-source-id: 49f0ec5407ea1f641896f054cad2283faed81687