port div to structured (#53680)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53680

Porting `div` to structured.

One weird thing to call out with div: It has an overload, `div.Tensor_mode`, which uses different TensorIterator settings depending on this input (the "mode" argument that you pass to it). So I ended up switching on the mode inside of the meta function to determine which TensorIterator builder to use.

Test Plan: Imported from OSS

Reviewed By: ezyang

Differential Revision: D27029819

Pulled By: bdhirsh

fbshipit-source-id: 3f216f6c197a2321087b4c23202bc2fc561491ba