[PyTorch Edge] Provide a method ObservedOperators::getUnobservedOperatorList() so that model tracer can empty it out during tracing (#55017)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55017

JacobSzwejbka in D26678637 found the mis-alignment between the operator list that the YAML file claimed and the dispatcher claimed. After some digging and thorough investigation by JacobSzwejbka, we have come to the conclusion that the non-traced operators are more trouble than they are worth since they will result in phantom operators which every user of the capabilities API needs to be aware of (or every language implementation needs to be aware of). Instead, with this change, we can **reliably** trace all operators called via the dispatcher by clearing the list of un-observed operators during model tracing.

Also another thing to note is that the ignore-list in the observer is a list of base operator names, and not full operator names (with overload), which is whaat tracing based selective build needs. If we use the ignore-list, then we would need to include every overload on un-traced operators.

Latency isn't an issue during model tracing, so this should be generally okay.

Ran the following command to re-generate all the YAML files: `buck run caffe2/torch/fb/mobile/cli:cli -- --gen_all_model_configs`
ghstack-source-id: 125337353

(Note: this ignores all push blocking failures!)

Test Plan: Sandcastle and wait for unit tests. Also see BSB results in the diff comments.

Reviewed By: JacobSzwejbka

Differential Revision: D27452855

fbshipit-source-id: 410bafec7ac67503f68623a5e3d4ab258f434cbf