Prepare for TensorPipe separating its CUDA-specific headers (#59788)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59788

This one line is all we need to "migrate" PyTorch to the "new API" of TensorPipe that splits the CUDA-specific stuff in a separate top-level header. (The motivation behind that is that it will allow us to "stack" the CUDA code on top of the CPU one).
ghstack-source-id: 131326166

Test Plan: None yet

Reviewed By: beauby

Differential Revision: D28875277

fbshipit-source-id: ecfd0b7fc0218ab7899bfe64ffe73c1417b897db