[package] minor cleanups to internal APIs (#61428)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61428

I was reading this code again after a while and didn't understand as
quickly as I would have liked. Some of the function names are no longer
accurate, etc.

This PR renames these functions to be in the same language of
"dependencies" that the rest of the API uses. I think the resulting
usage of the APIs is more clear than before

Test Plan: Imported from OSS

Reviewed By: Chillee

Differential Revision: D29620946

Pulled By: suo

fbshipit-source-id: 7df640a7ffbd43998063b9ee3955c9dfcbc42cfb