Relax use-count check to allow for 0 (#61414)
Summary:
Previously we require tensor use count to be exactly 1. We should actually allow for use count to be zero as well. Use count is zero when an undefined tensor is returned, and this is common in backward functions that have multiple outputs.

In this PR I also remove some entries from the skip list that should be covered by this change: they return multiple tensors AND are backward functions. Batch norm is also known to return undefined tensors when `training=False`.

Related issue: https://github.com/pytorch/pytorch/issues/60426

Pull Request resolved: https://github.com/pytorch/pytorch/pull/61414

Reviewed By: albanD

Differential Revision: D29614687

Pulled By: soulitzer

fbshipit-source-id: ab0892aed4bd1346b50b0a9552ffcc3287ac96af