SGX can not read from /dev/urandom (#60368)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60368

Problem:
The SGX secure enclave does not support reading from /dev/urandom as it is isolated from the OS for greater security. The SGX api provides a way to generate random numbers as a replacment.
Solution:
Conditionally enable SGX api for random number generation when building for it.

Test Plan: Run the PyTorch tests

Reviewed By: malfet, LiJihang

Differential Revision: D29022616

fbshipit-source-id: 1c7115457a2abde682df4d55fa4a8446fc5f8613