Remove OpInfo non-contig inputs (#67677)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67677

This follows
https://github.com/pytorch/pytorch/issues/63341#issuecomment-899690614

Fixes https://github.com/pytorch/pytorch/issues/67012

Note. I wrote the OpInfo for `index_fill`, so removing those inputs in
there is right. kshitij12345 mentioned that the same thing is true for
the inputs for tile / repeat.
https://github.com/pytorch/pytorch/issues/67012#issuecomment-948537446

There are more uses of `transpose` within the OpInfos, but most of them
are for testing `mm` and `baddmm`. I did not touch those, as those
operations are so important that it won't hurt to test those more
thoroughly.

cc mruberry

Test Plan: Imported from OSS

Reviewed By: malfet

Differential Revision: D32311729

Pulled By: mruberry

fbshipit-source-id: ac0804ca6f893118046b3e1bd97b5a2e6b900b59