[jit][edge] Return a no-op nullptr for UnionType on mobile for backward compatibility. (#71341)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71341

Old models containing UnionType need to be loaded even if they don't actually use Unions.
This is not the best solution, we need to catch this error on the compiler side instead, but before doing that we can land this first to at least mitigate model loading crash issues.
ghstack-source-id: 147056684

Test Plan:
CI
Verified with jaebong on his device locally.

Differential Revision: D33593276

fbshipit-source-id: fac4bc85c652974c7c10186a29f36e3e411865ad