[PyTorch][Lazy] Make hashing null optionals cheap (#71290)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71290

The existing code called an out-of-line hash function on a constant. This is just going to get the same random-looking 64-bit integer every time, so I just changed the constant to an integer I generated with `hex(random.randint(0x1000000000000000, 0xFFFFFFFFFFFFFFFF))` to get the same effect but without the runtime hashing.
ghstack-source-id: 146991945

Test Plan: CI

Reviewed By: wconstab

Differential Revision: D33574676

fbshipit-source-id: d6ce1e1cc0db67dfede148b7e3173508ec311ea8