Add knobs for XLA build options (#77729)
This is to enable XLA build option controls for cloud cache, and bazel sandboxed strategies. Some of the configurations worked in the downstream, but crashed the upstream XLA workflows. This doesn't change the current build configurations in the XLA workflow, but helps us decouple/configure different build options for the downstream CI runs without affecting the upstream.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77729
Approved by: https://github.com/seemethere