[ao] Adding model report detector base class and implemented detectors
Summary: the goal is to add a base class to the model report detectors
so that they can contain a lot more specific information compared to the
primary model report class related to the observers and where they are
inserted etc.

Since this is just a base class, the testing will be with the
implemenations of the classes that derive from the base class

The two current detector methods were turned into Detector classes and
the tests were modified to reflect this, but the same functionality was
tested.

As a result, _detector.py was changed to detector.py

Test Plan: python test/test_quantization.py TestFxModelReportDetector
python test/test_quantization.py TestFxModelReportDetectDynamicStatic

Reviewers:

Subscribers:

Tasks:

Tags:

Pull Request resolved: https://github.com/pytorch/pytorch/pull/79671

Approved by: https://github.com/andrewor14