[Profiler] Add pattern to detect if TF32 is available but not used (#81273)
Summary: This commit adds recording of a global flag allow_tf32 in all torch ops. Then I add a pattern that detects if a user is setting tf32 flag or not, recommend he/she to do so if is not set.

Test Plan: Added a test in test_profiler.py

Differential Revision: [D37799414](https://our.internmc.facebook.com/intern/diff/D37799414)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81273
Approved by: https://github.com/robieta