Add underlying_store property for PrefixStore (#84640)
Add a property to `PrefixStore` to retrieve the underlying store it is wrapping around. Open for suggestions on property name. This change is based on discussion in [D39225101](https://www.internalfb.com/diff/D39225101) where we need to read properties of the store that PrefixStore is wrapping around.

Differential Revision: [D39311151](https://our.internmc.facebook.com/intern/diff/D39311151)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84640
Approved by: https://github.com/xush6528