[exir] Add deepcopy to ExportedProgram (#105852)
Summary: ExirExportedProgram would like to have this feature. Today it does it itself since it inherits from ExportedProgram but since we are moving it to composition I think it would be cleaner to upstream the behavior into the root object anyway

Test Plan: ci, but todo where are the tests for this file?

Differential Revision: D47645843

Pull Request resolved: https://github.com/pytorch/pytorch/pull/105852
Approved by: https://github.com/tugsbayasgalan