[inductor] Handle more edge cases in slice and slice_scatter (#117377)
Fixes #117110

When slicing we can end up with start and end which are out of bounds, which is
handled in python slicing by clamping to the correct bounds. There is also the
case where end < start which should result in an empty slice.

In the isoneutral_mixing failure we have the second case, with `start=2, end=0`
which in `slice_scatter` became `src_size[dim] = -2`.

This PR improves slice's edge case handling and factors the start and end
normalization code out so it can be shared with slice_scatter.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/117377
Approved by: https://github.com/lezcano